commit para nuevas funciones

+++ golite/src/main/java/olc1/golite/visitor/interpreter/value/DecimalValue.java
@@ -4,7 +4,7 @@ public record DecimalValue(double value, int line, int column) implements ValueW
     
     @Override
     public String getTypeName() {
-        return "decimal";
+        return "float64";
     }
 
     @Override
